Seperated Category page controller from Session Controller
