+++ app/controllers/SessionsController.scala
@@ -15,7 +15,7 @@ import play.api.data.Forms._
 import play.api.data._
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.libs.json.{Json, OFormat}
-import play.api.mvc.{Action, AnyContent, Result}
+import play.api.mvc.{Action, AnyContent}
 import reactivemongo.bson.BSONDateTime
 import utilities.DateTimeUtility
 
@@ -66,7 +66,6 @@ case class KnolxSession(id: String,
                         expired: Boolean = false)
 
 case class SessionEmailInformation(email: Option[String], page: Int)
-case class ModelsCategoryInformation(_id: String, categoryName: String, subCategory: List[String])
 case class SessionSearchResult(sessions: List[KnolxSession],
                                pages: Int,
                                page: Int,
@@ -448,250 +447,6 @@ class SessionsController @Inject()(messagesApi: MessagesApi,
       eventualMaybeSession.flatMap(maybeSession =>
         maybeSession.fold(Future.successful(Redirect(routes.SessionsController.sessions(1, None)).flashing("message" -> "Session Not Found")))
         (session => Future.successful(Ok(views.html.sessions.sessioncontent(session)))))
-    }
-  }
-
-  def renderCategoryPage: Action[AnyContent] = adminAction.async { implicit request =>
-    Logger.info("-------------------------render category Page")
-    categoriesRepository.getCategories.map {
-      Logger.info("-------------------------Inside render category page")
-      category =>
-        Logger.info("--------------------------Caetgory List " + category)
-        Ok(views.html.category(category))
-    }
-  }
-
-  def addPrimaryCategory(categoryName: String): Action[AnyContent] = superUserAction.async { implicit request =>
-    Logger.info("Inside add primary category")
-    if (categoryName.trim().isEmpty) {
-      Future.successful(BadRequest("Primary category cannot be empty"))
-    } else {
-      categoriesRepository.getCategories.flatMap { result =>
-        if (result.exists(category => category.categoryName.toLowerCase.equals(categoryName.toLowerCase))) {
-          Future.successful(BadRequest("Primary category already exits"))
-        } else {
-          categoriesRepository.insertCategory(categoryName).map { result =>
-            if (result.ok) {
-              Ok("Primary category was successfully added")
-            } else {
-              BadRequest("Primary category cannot be added due to some error")
-            }
-          }
-        }
-      }
-    }
-  }
-
-  def addSubCategory(categoryName: String, subCategory: String): Action[AnyContent] = adminAction.async { implicit request =>
-    Logger.info("Inside add Sub category")
-    if (subCategory.trim().isEmpty) {
-      Future.successful(BadRequest("Subcategory cannot be empty"))
-    } else {
-      categoriesRepository.getCategories.flatMap { result =>
-        result.find {
-          _.categoryName == categoryName
-        }
-          .fold {
-            Future.successful(BadRequest("No primary category found."))
-          } { categoryInfo =>
-            val newSubCategory = categoryInfo.subCategory.find(_.toLowerCase == subCategory.toLowerCase)
-            newSubCategory.fold {
-              val subCategoryInfo = CategoryInfo(categoryName, List(subCategory), categoryInfo._id)
-              Logger.error("------>" + subCategoryInfo + "-------------->" + categoryInfo)
-              categoriesRepository.upsert(subCategoryInfo).map { result =>
-
-                Logger.info("Inside add sub category1 " + result)
-                if (result.ok) {
-                  Ok("Sub-category was successfully added")
-                } else {
-                  Logger.info("Inside bad request")
-                  BadRequest("Unsuccessful sub-category added")
-                }
-              }
-            } { _ =>
-              Future.successful(BadRequest("Error ! Sub-category already exist"))
-            }
-          }
-      }
-    }
-  }
-
-  def modifyPrimaryCategory(categoryId: String, newCategoryName: String): Action[AnyContent] = superUserAction.async { implicit request =>
-    Logger.info("Inside modify primary category")
-    if (newCategoryName.trim().isEmpty) {
-      Future.successful(BadRequest("Modify primary category cannot be empty"))
-    } else {
-      Logger.info("Old category name = " + categoryId)
-      Logger.info("new category name = " + newCategoryName)
-      categoriesRepository.getCategories.flatMap { result =>
-        val category = result.filter(c => c._id.stringify == categoryId).head
-        sessionsRepository.updateCategoryOnChange(category.categoryName, newCategoryName).flatMap { session =>
-          if (session.ok) {
-            categoriesRepository.modifyPrimaryCategory(category._id.stringify, newCategoryName).map { result =>
-              if (result.ok) {
-                Ok("Primary category was successfully modified")
-              } else {
-                Logger.info("Error Inside Sessions Controller")
-                BadRequest("Unsuccessfully attempt to modify primary category")
-              }
-            }
-          } else {
-            Future.successful(BadRequest("Update on session table was unsuccessful"))
-          }
-        }
-      }
-    }
-  }
-
-  def modifySubCategory(categoryName: String, oldSubCategoryName: String,
-                        newSubCategoryName: String): Action[AnyContent] = adminAction.async { implicit request =>
-    Logger.info("----------------------Inside modifySubCategory")
-    if (newSubCategoryName.trim().isEmpty) {
-      Future.successful(BadRequest("Modify sub-category cannot be empty"))
-    } else {
-      Logger.info("Before modifying sub category in session table")
-      categoriesRepository.getCategories.flatMap {
-        categories =>
-          val subCategoryList = categories.filter {
-            category => category.categoryName.toLowerCase == categoryName.toLowerCase
-          }.flatMap(_.subCategory)
-
-          val check = subCategoryList.contains(newSubCategoryName)
-          Logger.error("list--->" + subCategoryList + "new ------>  " + newSubCategoryName.toLowerCase + "---check-->  " + check)
-          if (check) {
-            Future.successful(BadRequest("Sub category already exits"))
-          } else {
-            sessionsRepository.updateSubCategoryOnChange(oldSubCategoryName, newSubCategoryName).flatMap { session =>
-              Logger.info("Modifying sub category in session table ")
-              if (session.ok) {
-                categoriesRepository.modifySubCategory(categoryName, oldSubCategoryName, newSubCategoryName).map {
-                  result =>
-                    if (result.ok) {
-                      Logger.info("--------------------------- Returning Ok")
-                      Ok("Successfully Modified sub category")
-                    } else {
-                      Logger.info("--------------------------- Returning Bad Request")
-                      BadRequest("Got an error while modifying sub category")
-                    }
-                }
-              } else {
-                Future.successful(BadRequest("Got an error on updating session table"))
-              }
-            }
-          }
-      }
-    }
-  }
-
-  def deletePrimaryCategory(categoryId: String): Action[AnyContent] = superUserAction.async { implicit request =>
-    Logger.info("Inside delete primary category" + categoryId)
-    if (categoryId.trim().isEmpty) {
-      Future.successful(BadRequest("Please select a valid primary category"))
-    } else {
-      categoriesRepository.getCategories.flatMap {
-        categories =>
-          Logger.info("List of categories = " + categories)
-          Logger.info("CategoryId = " + categoryId)
-          val category = categories.find {
-            c =>
-              Logger.info("current category id is = " + c._id + " and is being compared with categoryId" + categoryId)
-              c._id.toString() == categoryId
-          }
-          category.fold {
-            Future.successful(BadRequest("No such primary category exists"))
-          } { category =>
-            val subCategoryList = category.subCategory
-            Logger.info("Primary category with its sub category " + subCategoryList)
-            if (subCategoryList.isEmpty) {
-              Logger.info("Update Session Table on category delete")
-              sessionsRepository.updateCategoryOnChange(category.categoryName, "").flatMap { session =>
-                if (session.ok) {
-                  Logger.info("Delete primary category on delete")
-                  categoriesRepository.deletePrimaryCategory(category._id.stringify).map { result =>
-                    if (result.ok) {
-                      Ok("Primary category was successfully deleted")
-                    } else {
-                      BadRequest("Got an error while deleting")
-                    }
-                  }
-                } else {
-                  Future.successful(BadRequest("Got an error in session table"))
-                }
-              }
-            } else {
-              Future.successful(BadRequest("First delete all its sub category"))
-            }
-          }
-      }
-    }
-  }
-
-  def getSubCategoryByPrimaryCategory(categoryName: String): Action[AnyContent] = superUserAction.async { implicit request =>
-    if (categoryName.trim().isEmpty) {
-      Future.successful(BadRequest("Please select a valid primary category"))
-    } else {
-      categoriesRepository.getCategories.map {
-        categories =>
-          val subCategoryList = categories.filter {
-            category => category.categoryName == categoryName
-          }.flatMap(_.subCategory)
-          Logger.info("Primary category with its sub category " + subCategoryList)
-
-          if (subCategoryList.isEmpty) {
-            BadRequest(Json.toJson(subCategoryList).toString)
-          } else {
-            Ok(Json.toJson(subCategoryList).toString())
-          }
-      }
-    }
-  }
-
-  def deleteSubCategory(categoryName: String, subCategory: String): Action[AnyContent] = adminAction.async { implicit request =>
-    if (subCategory.trim.isEmpty) {
-      Future.successful(BadRequest("Sub-category cannot be empty"))
-    } else {
-      Logger.info(s"..........Before session repository $categoryName $subCategory")
-      sessionsRepository.updateSubCategoryOnChange(subCategory, "").flatMap { sessions =>
-        Logger.info("Inside session repository")
-        if (sessions.ok) {
-          Logger.info(".............. Before categories repository")
-          categoriesRepository.deleteSubCategory(categoryName, subCategory).flatMap { result =>
-            if (result.ok) {
-              Future.successful(Ok("Sub-category was successfully deleted"))
-            } else {
-              Future.successful(BadRequest("Something went wrong! unable to delete category"))
-            }
-          }
-        } else {
-          Future.successful(BadRequest("Got an error while deleting"))
-        }
-      }
-    }
-  }
-
-  def getTopicsBySubCategory(categoryName: String, subCategory: String): Action[AnyContent] = adminAction.async { implicit request =>
-    if (subCategory.trim.isEmpty) {
-      Future.successful(BadRequest("Please select a valid sub-category"))
-    } else {
-      Logger.info("category name = " + categoryName)
-      Logger.info("Sub category name =" + subCategory)
-      sessionsRepository.sessions.map { sessionInformation =>
-        val sessionTopicList = sessionInformation.filter {
-          session => session.subCategory == subCategory && session.category == categoryName
-        }.map {
-          _.topic
-        }
-        Logger.info("Inside get topics by subCategory = " + sessionTopicList)
-        Ok(Json.toJson(sessionTopicList).toString())
-      }
-    }
-  }
-
-  def getCategory: Action[AnyContent] = action.async { implicit request =>
-    Logger.info("Inside get category of controller")
-    categoriesRepository.getCategories.map { categories =>
-      val listOfCategoryInfo = categories.map(category => ModelsCategoryInformation(category._id.stringify, category.categoryName, category.subCategory))
-      Ok(Json.toJson(listOfCategoryInfo).toString)
     }
   }
