+++ app/views/header.scala.html
@@ -109,8 +109,8 @@
                             <i class="fa fa-folder-open" aria-hidden="true"></i>
                             <span>Reports</span></a>
                     </li>
-                    <li class="@activeLink(routes.SessionsController.renderCategoryPage().url) alignLinks" data-toggle="tooltip" data-placement="right" title="Category">
-                        <a href="@routes.SessionsController.renderCategoryPage()">
+                    <li class="@activeLink(routes.SessionsCategoryController.renderCategoryPage().url) alignLinks" data-toggle="tooltip" data-placement="right" title="Category">
+                        <a href="@routes.SessionsCategoryController.renderCategoryPage()">
                             <i class="fa fa-plus" aria-hidden="true"></i>
                             <span>Category</span></a>
                     </li>
