--- app/views/header2.scala.html
@@ -1,123 +1,0 @@
-@()(implicit request: Request[_])
-
-@activeLink(currentPath: String) = @{
-if(request.path.equals(currentPath)) "active"
-}
-
-<div class="wrapper">
-
-        <!-- Sidebar Holder -->
-    <nav class="navbar navbar-fixed-top" role="navigation" style="margin-bottom: 0">
-        <div class="navbar-header" style="height: 50px;">
-            <span class="main-logo" >
-                <a href="#" class="pull-left logo">
-                    <img src='@routes.Assets.versioned("images/logo3.png")' style="padding-left: 15px;">
-                </a>
-            </span>
-        </div>
-        @*<a id="sidebarCollapse" href="#" data-side="mini" style="padding: 16px 15px; border-right: 1px solid #e7e7e7 !important; height: 50px;">
-            <i class="fa fa-fw fa-bars" style="margin-top: 18px;"></i>
-        </a>*@
-
-        <ul class="nav navbar-top-links navbar-right" style="margin:1px;">
-        @if(SessionHelper.isLoggedIn) {
-            <li class="@activeLink(routes.UsersController.login().url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Login">
-                <a href="@routes.UsersController.login()"><i class="fa fa-sign-in"></i>
-                    <span>Login</span></a>
-            </li>
-            <li class="@activeLink(routes.UsersController.register().url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Register">
-                <a href="@routes.UsersController.register()"><i class="fa fa-user-plus"></i>
-                    <span>Register</span></a>
-            </li>
-        } else {
-            <li class="dropdown alignLinks" data-toggle="tooltip" data-placement="bottom" title="@SessionHelper.email.split('@')(0).split('.')(0)">
-                <a href="#" class="dropdown-toggle" data-toggle="dropdown">
-                    <i class="fa fa-user-circle-o"></i> <span>@SessionHelper.email.split("@")(0).split('.')(0)
-                </span>
-                    <b class="caret"></b>
-                </a>
-                <ul class="dropdown-menu" >
-                    <li class="alignLinks" data-toggle="tooltip" data-placement="bottom" title="Reset Password">
-                        <a href="@routes.UsersController.renderChangePassword()"><i class="fa fa-key" aria-hidden="true"></i>
-                            <span>Reset Password</span></a>
-                    </li>
-                    <li class="alignLinks" data-toggle="tooltip" data-placement="bottom" title="Logout">
-                        <a href="@routes.UsersController.logout()"><i class="fa fa-sign-out" aria-hidden="true"></i>
-                            Logout</a>
-                    </li>
-                </ul>
-            </li>
-
-        }
-
-        </ul>
-    </nav>
-    <nav id="sidebar">
-
-        <ul class="list-unstyled components" style="border-bottom: none;">
-            <li class="@activeLink("/session") alignLinks" data-toggle="tooltip" data-placement="bottom" title="Session">
-                <a href="@routes.SessionsController.sessions(1, None)">
-                    <i class="fa fa-home"></i>
-                    <span>Session</span>
-                </a>
-            </li>
-
-            <li class="@activeLink(routes.KnolxAnalysisController.renderAnalysisPage().url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Analysis">
-                <a href="@routes.KnolxAnalysisController.renderAnalysisPage"><i class="fa fa-line-chart"></i>
-                    <span>Analysis</span></a>
-            </li>
-
-            @if(!SessionHelper.isLoggedIn) {
-
-
-            <li class="@activeLink(routes.FeedbackFormsResponseController.getFeedbackFormsForToday().url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Feedback">
-                <a href="@routes.FeedbackFormsResponseController.getFeedbackFormsForToday"><i class="fa fa-comments"></i>
-                    <span>Feedbacks</span></a>
-            </li>
-            <li class="@activeLink(routes.FeedbackFormsReportController.renderUserFeedbackReports().url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Session Report">
-                <a href="@routes.FeedbackFormsReportController.renderUserFeedbackReports()"><i class="fa fa-table" aria-hidden="true"></i>
-                    <span>Knolx reports</span></a>
-            </li>
-            }
-
-            @if(SessionHelper.isAdmin) {
-
-            <li class="alignLinks" data-toggle="tooltip" data-placement="bottom" title="Manage">
-                <a href="#manageSubmenu" data-toggle="collapse" aria-expanded="false">
-                    <i class="fa fa-wrench"></i><span> Manage</span>
-                    <b class="caret"></b>
-                </a>
-
-                <ul class="collapse list-unstyled" id="manageSubmenu">
-                    <li class="@activeLink(routes.SessionsController.manageSessions(1, None).url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Sessions">
-                        <a href="@routes.SessionsController.manageSessions(1, None)"><i class="fa fa-hourglass-end" aria-hidden="true"></i>
-                            <span>Sessions</span></a>
-                    </li>
-                    <li class="@activeLink(routes.FeedbackFormsController.manageFeedbackForm(1).url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Feedback Forms">
-                        <a href="@routes.FeedbackFormsController.manageFeedbackForm(1)"><i class="fa fa-file" aria-hidden="true"></i>
-                            <span>Feedback Forms</span></a>
-                    </li>
-                    <li class="@activeLink(routes.UsersController.manageUser(1, None).url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Users">
-                        <a href="@routes.UsersController.manageUser(1, None)"><i class="fa fa-users" aria-hidden="true"></i>
-                            <span>Users</span></a>
-                    </li>
-                    <li class="@activeLink(routes.FeedbackFormsReportController.renderAllFeedbackReports().url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Reports">
-                        <a href="@routes.FeedbackFormsReportController.renderAllFeedbackReports()"><i class="fa fa-folder-open" aria-hidden="true"></i>
-                            <span>Reports</span></a>
-                    </li>
-                    <li class="@activeLink(routes.SessionsController.renderCategoryPage().url) alignLinks" data-toggle="tooltip" data-placement="bottom" title="Reports">
-                        <a href="@routes.SessionsController.renderCategoryPage()"><i class="fa fa-plus" aria-hidden="true"></i>
-                            <span>Category</span></a>
-                    </li>
-                </ul>
-            </li>
-
-            }
-        </ul>
-        <div id="sidebarCollapse">
-            <i id="collapse-button" class="fa fa-angle-double-left">
-            </i>
-        </div>
-    </nav>
-    <div id="content" style="margin-top: 60px;">
-        <div>
