+++ app/views/main.scala.html
@@ -3,27 +3,27 @@
 <!DOCTYPE html>
 <html lang="en">
 
-    <head>
-        <title>@title</title>
+<head>
+    <title>@title</title>
 
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width; initial-scale=0.0; user-scalable=0">
-            @*knolx-stylesheets*@
-            @knolxstylesheet()
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width; initial-scale=0.0; user-scalable=0">
+    @*knolx-stylesheets*@
+    @knolxstylesheet()
 
-            @*knolx- javascripts*@
-            @knolxjavascripts()
-    </head>
-    <body>
-        @header()
-    <div class="container-knolx">
+    @*knolx- javascripts*@
+    @knolxjavascripts()
+</head>
+<body>
+@header()
+<div class="container-knolx">
     @content
-    </div>
-        @footer()
-    </div>
-    </div>
-    <script src="@routes.Assets.versioned("javascripts/loader.js")" type="text/javascript"></script>
+</div>
+@footer()
+</div>
+</div>
+<script src="@routes.Assets.versioned("javascripts/loader.js")" type="text/javascript"></script>
 
     </body>
